backend/U2net/U-2-Net/requirements.txt
flask
flask-restful
flask-jwt-extended
pika  
python-dotenv 
torchvision
torchvision 
numpy 
Pillow 
scikit-image
pika
matplotlib